requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
et_xmlfile==2.0.0
greenlet==3.1.1
idna==3.10
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
pyTelegramBotAPI==4.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.32.3
six==1.17.0
SQLAlchemy==1.4.46
telebot==0.0.5
tzdata==2025.2
urllib3==2.4.0
